DLL 関係の処理を移動
