+++ daemon/Cargo.toml
@@ -10,6 +10,7 @@ license = "AGPL-3.0-or-later"
 [dependencies]
 aquestalk-proxy = { path = "../lib" }
 getopts = "0.2.21"
+libloading = "0.7"
 optional_take = "0.1.0"
 serde_json = "1.0"
 threadpool = "1.8"
