+++ daemon/src/aquestalk/dll.rs
@@ -1,10 +1,19 @@
-// Copyright (c) 2021-2022 Na-x4
+// AquesTalk-proxy - Copyright (C) 2021-2022 Na-x4
 //
-// Licensed under the Apache License, Version 2.0 <LICENSE-APACHE or
-// https://www.apache.org/licenses/LICENSE-2.0> or the MIT license
-// <LICENSE-MIT or https://opensource.org/licenses/MIT>, at your
-// option. This file may not be copied, modified, or distributed
-// except according to those terms.
+// This file is part of AquesTalk-proxy.
+//
+// AquesTalk-proxy is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// AquesTalk-proxy is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with AquesTalk-proxy.  If not, see <https://www.gnu.org/licenses/>.
 
 use std::ffi::OsStr;
 use std::os::raw::{c_char, c_int, c_uchar};
