+++ lib/Cargo.toml
@@ -10,6 +10,5 @@ license = "MIT OR Apache-2.0"
 [dependencies]
 base64 = "0.13"
 encoding_rs = "0.8"
-libloading = "0.7"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
